renamed controllers to routes
